requirements.txt
flask
requests
gunicorn
python-dotenv
aiohttp
